Inertia View path might be different
The view path for Inertia pages might be different in some projects. The default path according to the Inertia documentation is `ressources/js/Pages`. However as a developer I could change it to whatever I like. Even moving the Pages folder to a completely different place.

```js
createInertiaApp({
  resolve: name => require(`./Madness/${name}`),
  // ...
})
```

Therefore it is not possible for Laravel to check if a component exists.